Create a Government_type_id, following the pattern of Impr_type_id etc.
Get rid of G_MAGIC and the (unused) G_CITY_SIZE_FREE.

See gna patch #7277.

[[originally from svn r32924]]

+++ server/report.c
@@ -766,7 +766,7 @@ static int get_specialists(const struct player *pplayer)
 ****************************************************************************/
 static int get_gov(const struct player *pplayer)
 {
-  return government_number(government_of_player(pplayer));
+  return (int) government_number(government_of_player(pplayer));
 }
 
 /****************************************************************************
